refactor(contributors): split fetch logic into model layer and move CLI to bin/
