+++ roquiz/lib/model/persistence/contributors_repository.dart
@@ -3,7 +3,7 @@ import 'package:roquiz/model/persistence/contributor_info.dart';
 import 'package:yaml/yaml.dart';
 
 /// Loads the contributors list bundled with the app. The YAML is generated by
-/// `lib/cli/fetch_contributors.dart`.
+/// the `fetch_contributors` CLI tool (`bin/fetch_contributors.dart`).
 class ContributorsRepository {
   static const String defaultAssetPath = "assets/contributors/contributors.yaml";
